Refactoring strategies (#13)
* Move strategies to a separate package

* Pass the Deck to the GameEngine

* Pass the Deck as an argument to the TurnStrategy

The Deck is not a dependency for a TurnStrategy but is needed in the context of a playTurn method. Therefore the TurnStrategy can be reused across games and does not need a Deck in advance.

* Shuffle the deck in the GameEngine

Each game is responsible for shuffling its Deck before starting. This is now tested too.

* Reuse strategies across games